move `requireElement` to `player-api.ts`
